healer

+++ Source/TheEarlyManWars/Assets/Scripts/Towers/TowerObject.cs
@@ -5,6 +5,6 @@ public class TowerObject : ScriptableObject
     public int level;
     public int hp;
     public int rangeAttack;
-    public int attackSpeed;
+    public float attackSpeed;
     public TowerDisplay displayPrefab;
 }
